requirements.txt
channels==4.1.0
daphne==4.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-jazzmin==2.6.1
djangorestframework==3.15.1
drf-yasg==1.21.7
Pillow==10.3.0
pytz==2024.1
rel==0.4.9.9
websocket-client==1.7.0